Read and print the output from 'extended_summary'

extended_summary reads and prints the saved output from 'extended_summary'.

read_summary(path = ".", fill = "#6DCD59FF", highlight = FALSE)

Arguments

path

Directory in which to look for the 'extended_summary' output files.

fill

Color for posterior distribution fill. Default is "#6DCD59FF".

highlight

Logical to control if posterior estimates for which the 95% credible intervals do not overlap with zero are highlighted. Default is FALSE.

Value

If plot = TRUE the function reads the saved output of extended_summary and prints it.

Details

It prints a summary of brmsfit results that were previously saved by extended_summary. Useful when models are too big so reading them many times is to be avoided. It includes a model fit table, a coefficient table and a posterior distribution halfeye graph next to a chain trace plot. Tables are produce in html format so the output is nicely printed when knit in Rmarkdown files. You might have to add 'results = 'as.is', warning = FALSE and adjust fig.width to chunk options in Rmarkdown documents.

References

Araya-Salas (2022), brmsish: miscellaneous functions to customize brms bayesian regression models. R package version 1.0.0.Paul-Christian Buerkner (2017). brms: An R Package for Bayesian Multilevel Models Using Stan. Journal of Statistical Software, 80(1), 1-28. doi:10.18637/jss.v080.i01

See also

Author

Marcelo Araya-Salas marcelo.araya@ucr.ac.cr)

Examples

{
# run model
md1 <- brm(Petal.Length ~ Petal.Width + Species, iris, chains = 1,
iter = 500)

# save summary
extended_summary(fit = md1, save = TRUE, dest.path = tempdir())

# read/print summary
read_summary(path = file.path(tempdir(), "md1"))
}
#> Compiling Stan program...
#> Start sampling
#> 
#> SAMPLING FOR MODEL '0b59cc4e3935f086e2dde5c84040063f' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.2e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.12 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:   1 / 500 [  0%]  (Warmup)
#> Chain 1: Iteration:  50 / 500 [ 10%]  (Warmup)
#> Chain 1: Iteration: 100 / 500 [ 20%]  (Warmup)
#> Chain 1: Iteration: 150 / 500 [ 30%]  (Warmup)
#> Chain 1: Iteration: 200 / 500 [ 40%]  (Warmup)
#> Chain 1: Iteration: 250 / 500 [ 50%]  (Warmup)
#> Chain 1: Iteration: 251 / 500 [ 50%]  (Sampling)
#> Chain 1: Iteration: 300 / 500 [ 60%]  (Sampling)
#> Chain 1: Iteration: 350 / 500 [ 70%]  (Sampling)
#> Chain 1: Iteration: 400 / 500 [ 80%]  (Sampling)
#> Chain 1: Iteration: 450 / 500 [ 90%]  (Sampling)
#> Chain 1: Iteration: 500 / 500 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.019486 seconds (Warm-up)
#> Chain 1:                0.013809 seconds (Sampling)
#> Chain 1:                0.033295 seconds (Total)
#> Chain 1: 
#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#bulk-ess
#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#tail-ess
#> 
#> 
#> ##  md1 
#> 
#> <table class="table table-striped table-hover table-condensed table-responsive" style="font-size: 12px; width: auto !important; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;">   </th>
#>    <th style="text-align:left;"> priors </th>
#>    <th style="text-align:left;"> formula </th>
#>    <th style="text-align:right;"> iterations </th>
#>    <th style="text-align:right;"> chains </th>
#>    <th style="text-align:right;"> thinning </th>
#>    <th style="text-align:right;"> warmup </th>
#>    <th style="text-align:left;"> diverg_transitions </th>
#>    <th style="text-align:left;"> rhats &gt; 1.05 </th>
#>    <th style="text-align:right;"> min_bulk_ESS </th>
#>    <th style="text-align:right;"> min_tail_ESS </th>
#>    <th style="text-align:right;"> seed </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> 1 </td>
#>    <td style="text-align:left;"> Intercept-student_t(3, 4.3, 2.5)
#> sigma-student_t(3, 0, 2.5) </td>
#>    <td style="text-align:left;"> Petal.Length ~ Petal.Width + Species </td>
#>    <td style="text-align:right;"> 500 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 250 </td>
#>    <td style="text-align:left;"> <span style="     ">0</span> </td>
#>    <td style="text-align:left;"> <span style="     ">0</span> </td>
#>    <td style="text-align:right;"> 86.884 </td>
#>    <td style="text-align:right;"> 65.115 </td>
#>    <td style="text-align:right;"> 1190607795 </td>
#>   </tr>
#> </tbody>
#> </table><table class="table table-striped table-hover table-condensed table-responsive" style="font-size: 12px; width: auto !important; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;">   </th>
#>    <th style="text-align:right;"> Estimate </th>
#>    <th style="text-align:right;"> l-95% CI </th>
#>    <th style="text-align:right;"> u-95% CI </th>
#>    <th style="text-align:left;"> Rhat </th>
#>    <th style="text-align:right;"> Bulk_ESS </th>
#>    <th style="text-align:right;"> Tail_ESS </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> b_Intercept </td>
#>    <td style="text-align:right;"> 1.210 </td>
#>    <td style="text-align:right;"> 1.078 </td>
#>    <td style="text-align:right;"> 1.350 </td>
#>    <td style="text-align:left;"> <span style="     ">1</span> </td>
#>    <td style="text-align:right;"> 231.229 </td>
#>    <td style="text-align:right;"> 189.326 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> b_Petal.Width </td>
#>    <td style="text-align:right;"> 1.006 </td>
#>    <td style="text-align:right;"> 0.742 </td>
#>    <td style="text-align:right;"> 1.344 </td>
#>    <td style="text-align:left;"> <span style="     ">1.004</span> </td>
#>    <td style="text-align:right;"> 87.896 </td>
#>    <td style="text-align:right;"> 65.115 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> b_Speciesversicolor </td>
#>    <td style="text-align:right;"> 1.709 </td>
#>    <td style="text-align:right;"> 1.277 </td>
#>    <td style="text-align:right;"> 2.035 </td>
#>    <td style="text-align:left;"> <span style="     ">0.996</span> </td>
#>    <td style="text-align:right;"> 91.360 </td>
#>    <td style="text-align:right;"> 67.422 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> b_Speciesvirginica </td>
#>    <td style="text-align:right;"> 2.300 </td>
#>    <td style="text-align:right;"> 1.666 </td>
#>    <td style="text-align:right;"> 2.830 </td>
#>    <td style="text-align:left;"> <span style="     ">0.997</span> </td>
#>    <td style="text-align:right;"> 86.884 </td>
#>    <td style="text-align:right;"> 67.527 </td>
#>   </tr>
#> </tbody>
#> </table>![](/tmp/RtmpO8PbRq/md1/plot.jpeg)