Check brmsfit objects save as RDS files

check_rds_fits checks brmsfit objects save as RDS files.

check_rds_fits(path = ".", fits = list.files(path = path, pattern = ".RDS$",
ignore.case = TRUE, full.names = TRUE), cores = 1, pb = TRUE, robust = TRUE,
html = FALSE, verbose = TRUE)

Arguments

path

Directory in which to look for .rds files.

fits

Name of the .rds files to be read. Optional.

cores

Number of cores to use for parallelization. Default is 1 (no parallelization).

pb

Logical to control if a progress bar is used. Default is TRUE.

robust

Logical to control the type of central tendency measure as in summary.brmsfit).

html

Logical to control whether results are returned in html format. Useful for creating Rmd or quarto html reports. Is FALSE (default) the table is return as a data frame object.

verbose

Logical to control if messages are printed into the console.

Value

Returns a data frame with a summary of fitted models. Can be used to make sure all models were run with the same parameters (e.g. before combining models). If html = FALSE the function will return a data frame, otherwise it will print the estimates in a table in html format. The summary includes: prior, formula, number of iterations, number of chain, thinning, warmup, number of parameters, number of divergent transitions, number of rhats higher than 1.05, tail and bulk effective sample sizes and seed.

Details

The function reads all fits saved as rds files in the supplied directory and generates a table listing the parameters used to fit models.

References

Araya-Salas (2022), brmsish: miscellaneous functions to customize brms bayesian regression models. R package version 1.0.0.Paul-Christian Buerkner (2017). brms: An R Package for Bayesian Multilevel Models Using Stan. Journal of Statistical Software, 80(1), 1-28. doi:10.18637/jss.v080.i01

See also

Author

Marcelo Araya-Salas marcelo.araya@ucr.ac.cr)

Examples

{
# create directory
dir.create(file.path(tempdir(), "rdss"))
# run 2 models
md1 <- brm(Petal.Length ~ Petal.Width + Species, iris, chains = 1,
iter = 500, file = file.path(tempdir(), "rdss", "md1"))

md2 <- brm(Petal.Length ~ Species, iris, chains = 1,
iter = 500, file = file.path(tempdir(), "rdss", "md2"))

# check fits
check_rds_fits(path = file.path(tempdir(), "rdss"))
}
#> Compiling Stan program...
#> Start sampling
#> 
#> SAMPLING FOR MODEL '0b59cc4e3935f086e2dde5c84040063f' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.7e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.17 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:   1 / 500 [  0%]  (Warmup)
#> Chain 1: Iteration:  50 / 500 [ 10%]  (Warmup)
#> Chain 1: Iteration: 100 / 500 [ 20%]  (Warmup)
#> Chain 1: Iteration: 150 / 500 [ 30%]  (Warmup)
#> Chain 1: Iteration: 200 / 500 [ 40%]  (Warmup)
#> Chain 1: Iteration: 250 / 500 [ 50%]  (Warmup)
#> Chain 1: Iteration: 251 / 500 [ 50%]  (Sampling)
#> Chain 1: Iteration: 300 / 500 [ 60%]  (Sampling)
#> Chain 1: Iteration: 350 / 500 [ 70%]  (Sampling)
#> Chain 1: Iteration: 400 / 500 [ 80%]  (Sampling)
#> Chain 1: Iteration: 450 / 500 [ 90%]  (Sampling)
#> Chain 1: Iteration: 500 / 500 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.016843 seconds (Warm-up)
#> Chain 1:                0.019495 seconds (Sampling)
#> Chain 1:                0.036338 seconds (Total)
#> Chain 1: 
#> Warning: The largest R-hat is 1.08, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#r-hat
#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#bulk-ess
#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#tail-ess
#> Compiling Stan program...
#> recompiling to avoid crashing R session
#> Start sampling
#> 
#> SAMPLING FOR MODEL '0b59cc4e3935f086e2dde5c84040063f' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.3e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.13 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:   1 / 500 [  0%]  (Warmup)
#> Chain 1: Iteration:  50 / 500 [ 10%]  (Warmup)
#> Chain 1: Iteration: 100 / 500 [ 20%]  (Warmup)
#> Chain 1: Iteration: 150 / 500 [ 30%]  (Warmup)
#> Chain 1: Iteration: 200 / 500 [ 40%]  (Warmup)
#> Chain 1: Iteration: 250 / 500 [ 50%]  (Warmup)
#> Chain 1: Iteration: 251 / 500 [ 50%]  (Sampling)
#> Chain 1: Iteration: 300 / 500 [ 60%]  (Sampling)
#> Chain 1: Iteration: 350 / 500 [ 70%]  (Sampling)
#> Chain 1: Iteration: 400 / 500 [ 80%]  (Sampling)
#> Chain 1: Iteration: 450 / 500 [ 90%]  (Sampling)
#> Chain 1: Iteration: 500 / 500 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.007318 seconds (Warm-up)
#> Chain 1:                0.005752 seconds (Sampling)
#> Chain 1:                0.01307 seconds (Total)
#> Chain 1: 
#>       fit                                                       priors
#> 1 md1.rds Intercept-student_t(3, 4.3, 2.5)\nsigma-student_t(3, 0, 2.5)
#> 2 md2.rds Intercept-student_t(3, 4.3, 2.5)\nsigma-student_t(3, 0, 2.5)
#>                                formula iterations chains thinning warmup
#> 1 Petal.Length ~ Petal.Width + Species        500      1        1    250
#> 2               Petal.Length ~ Species        500      1        1    250
#>   n_parameters diverg_transitions rhats > 1.05 min_bulk_ESS min_tail_ESS
#> 1            7                  0            2     55.17232     31.70159
#> 2            6                  0            0    103.62701    102.99606
#>        seed
#> 1 259802396
#> 2 378562895